+++ src/app_info_ui.cpp
@@ -1,13 +1,16 @@
 #include "app_info_ui.hpp"
+#include "blacklist_settings_ui.hpp"
+#include "preset_type.hpp"
+#include "util.hpp"
 #include <glibmm/i18n.h>
 #include <sstream>
-#include "util.hpp"
 
 AppInfoUi::AppInfoUi(BaseObjectType* cobject,
                      const Glib::RefPtr<Gtk::Builder>& builder,
                      std::shared_ptr<AppInfo> info,
                      PulseManager* pulse_manager)
-    : Gtk::Grid(cobject), app_info(std::move(info)), pm(pulse_manager) {
+    : Gtk::Grid(cobject), app_info(std::move(info)), pm(pulse_manager),
+      settings(Gio::Settings::create("com.github.wwmm.pulseeffects")) {
   // loading glade widgets
 
   builder->get_widget("enable", enable);
@@ -15,6 +18,7 @@ AppInfoUi::AppInfoUi(BaseObjectType* cobject,
   builder->get_widget("app_name", app_name);
   builder->get_widget("volume", volume);
   builder->get_widget("mute", mute);
+  builder->get_widget("blacklist", blacklist);
   builder->get_widget("mute_icon", mute_icon);
   builder->get_widget("format", format);
   builder->get_widget("rate", rate);
@@ -23,6 +27,15 @@ AppInfoUi::AppInfoUi(BaseObjectType* cobject,
   builder->get_widget("buffer", buffer);
   builder->get_widget("latency", latency);
   builder->get_widget("state", state);
+
+  std::vector<std::string> blacklisted_apps =
+      settings->get_string_array((app_info->app_type == "sink_input") ? "blacklist-out" : "blacklist-in");
+  blacklisted =
+      std::find(std::begin(blacklisted_apps), std::end(blacklisted_apps), app_info->name) != std::end(blacklisted_apps);
+
+  enabled = app_info->connected && !blacklisted;
+  pre_bl_state =
+      settings->get_boolean((app_info->app_type == "sink_input") ? "enable-all-sinkinputs" : "enable-all-sourceoutputs");
 
   init_widgets();
   connect_signals();
@@ -60,7 +73,10 @@ auto AppInfoUi::latency_to_str(uint value) -> std::string {
 }
 
 void AppInfoUi::init_widgets() {
-  enable->set_active(app_info->connected);
+  enable->set_active(enabled && !blacklisted);
+  enable->set_sensitive(!blacklisted);
+
+  blacklist->set_active(blacklisted);
 
   app_icon->set_from_icon_name(app_info->icon_name, Gtk::ICON_SIZE_BUTTON);
 
@@ -95,6 +111,29 @@ void AppInfoUi::connect_signals() {
   volume_connection = volume->signal_value_changed().connect(sigc::mem_fun(*this, &AppInfoUi::on_volume_changed));
 
   mute_connection = mute->signal_toggled().connect(sigc::mem_fun(*this, &AppInfoUi::on_mute));
+
+  blacklist_connection = blacklist->signal_clicked().connect([=]() {
+    PresetType preset_type = (app_info->app_type == "sink_input") ? PresetType::output : PresetType::input;
+
+    if (blacklist->get_active()) {
+      // Add new entry to blacklist vector
+      BlacklistSettingsUi::add_new_entry(settings, app_info->name, preset_type);
+      pre_bl_state = enabled;
+      blacklisted = true;
+      if (enabled) {
+        enable->set_active(false);
+      }
+      enable->set_sensitive(false);
+    } else {
+      // Remove app name entry from blacklist vector
+      BlacklistSettingsUi::remove_entry(settings, app_info->name, preset_type);
+      blacklisted = false;
+      enable->set_sensitive(true);
+      if (pre_bl_state) {
+        enable->set_active(true);
+      }
+    }
+  });
 }
 
 auto AppInfoUi::on_enable_app(bool state) -> bool {
@@ -111,6 +150,8 @@ auto AppInfoUi::on_enable_app(bool state) -> bool {
       pm->remove_source_output_from_pulseeffects(app_info->name, app_info->index);
     }
   }
+
+  enabled = state;
 
   return false;
 }
@@ -151,6 +192,7 @@ void AppInfoUi::update(const std::shared_ptr<AppInfo>& info) {
   enable_connection.disconnect();
   volume_connection.disconnect();
   mute_connection.disconnect();
+  blacklist_connection.disconnect();
 
   init_widgets();
   connect_signals();