+++ src/sink_input_effects.cpp
@@ -5,7 +5,7 @@
 namespace {
 
 void on_message_element(const GstBus* gst_bus, GstMessage* message, SinkInputEffects* sie) {
-  auto src_name = GST_OBJECT_NAME(message->src);
+  auto* src_name = GST_OBJECT_NAME(message->src);
 
   if (std::strcmp(src_name, "pitch_input_level") == 0) {
     sie->pitch_input_level.emit(SinkInputEffects::get_peak(message));
@@ -74,7 +74,7 @@ SinkInputEffects::SinkInputEffects(PulseManager* pulse_manager) : PipelineBase("
   set_source_monitor_name(pm->apps_sink_info->monitor_source_name);
   set_caps(pm->apps_sink_info->rate);
 
-  auto PULSE_SINK = std::getenv("PULSE_SINK");
+  auto* PULSE_SINK = std::getenv("PULSE_SINK");
 
   if (PULSE_SINK != nullptr) {
     if (pm->get_sink_info(PULSE_SINK)) {
@@ -173,16 +173,35 @@ SinkInputEffects::~SinkInputEffects() {
 void SinkInputEffects::on_app_added(const std::shared_ptr<AppInfo>& app_info) {
   PipelineBase::on_app_added(app_info);
 
-  auto enable_all = g_settings_get_boolean(settings, "enable-all-sinkinputs");
+  bool forbidden_app = false;
+  auto* blacklist = g_settings_get_strv(settings, "blacklist-out");
 
-  if ((enable_all != 0) && !app_info->connected) {
-    pm->move_sink_input_to_pulseeffects(app_info->name, app_info->index);
+  for (std::size_t i = 0; blacklist[i] != nullptr; i++) {
+    if (app_info->name == blacklist[i]) {
+      forbidden_app = true;
+    }
+
+    g_free(blacklist[i]);
   }
+
+  if (app_info->connected) {
+    if (forbidden_app) {
+      pm->remove_sink_input_from_pulseeffects(app_info->name, app_info->index);
+    }
+  } else {
+    auto enable_all = g_settings_get_boolean(settings, "enable-all-sinkinputs");
+
+    if (!forbidden_app && enable_all != 0) {
+      pm->move_sink_input_to_pulseeffects(app_info->name, app_info->index);
+    }
+  }
+
+  g_free(blacklist);
 }
 
 void SinkInputEffects::add_plugins_to_pipeline() {
-  gchar* name;
-  GVariantIter* iter;
+  gchar* name = nullptr;
+  GVariantIter* iter = nullptr;
   std::vector<std::string> default_order;
 
   g_settings_get(child_settings, "plugins", "as", &iter);
@@ -192,7 +211,7 @@ void SinkInputEffects::add_plugins_to_pipeline() {
     g_free(name);
   }
 
-  auto gvariant = g_settings_get_default_value(child_settings, "plugins");
+  auto* gvariant = g_settings_get_default_value(child_settings, "plugins");
 
   g_variant_get(gvariant, "as", &iter);
 
@@ -227,7 +246,7 @@ void SinkInputEffects::add_plugins_to_pipeline() {
 
   // adding plugins to effects_bin
 
-  for (auto& p : plugins) {
+  for (const auto& p : plugins) {
     gst_bin_add(GST_BIN(effects_bin), p.second);
   }